loganalytics: splitting to support generating these
