+++ azurerm/internal/services/loganalytics/validate/log_analytics_cluster_name.go
@@ -1,0 +1,5 @@
+package validate
+
+func LogAnalyticsClusterName(i interface{}, k string) (warnings []string, errors []error) {
+	return logAnalyticsGenericName(i, k)
+}
