+++ azurerm/internal/services/loganalytics/validate/log_analytics_data_export_name.go
@@ -1,9 +1,5 @@
 package validate
 
-func LogAnalyticsDataExportWorkspaceName(i interface{}, k string) (warnings []string, errors []error) {
-	return logAnalyticsGenericName(i, k)
-}
-
 func LogAnalyticsDataExportName(i interface{}, k string) (warnings []string, errors []error) {
 	return logAnalyticsGenericName(i, k)
 }
