+++ azurerm/internal/services/loganalytics/validate/log_analytics_linked_storage_account_workspace_name.go
@@ -3,7 +3,3 @@ package validate
 func LogAnalyticsLinkedStorageAccountWorkspaceName(i interface{}, k string) (warnings []string, errors []error) {
 	return logAnalyticsGenericName(i, k)
 }
-
-func LogAnalyticsLinkedStorageAccountName(i interface{}, k string) (warnings []string, errors []error) {
-	return logAnalyticsGenericName(i, k)
-}
