+++ azurerm/internal/services/loganalytics/validate/log_analytics_linked_storage_account_workspace_name_test.go
