+++ azurerm/internal/services/loganalytics/validate/log_analytics_storage_insights_name.go
@@ -3,7 +3,3 @@ package validate
 func LogAnalyticsStorageInsightsName(i interface{}, k string) (warnings []string, errors []error) {
 	return logAnalyticsGenericName(i, k)
 }
-
-func LogAnalyticsStorageInsightsWorkspaceName(i interface{}, k string) (warnings []string, errors []error) {
-	return logAnalyticsGenericName(i, k)
-}
