+++ azurerm/internal/services/loganalytics/validate/log_analytics_storage_insights_workspace_name.go
@@ -1,0 +1,5 @@
+package validate
+
+func LogAnalyticsStorageInsightsWorkspaceName(i interface{}, k string) (warnings []string, errors []error) {
+	return logAnalyticsGenericName(i, k)
+}
