+++ azurerm/internal/services/loganalytics/validate/log_analytics_storage_insights_workspace_name_test.go
