+++ azurerm/internal/services/loganalytics/validate/log_analytics_workspace_id.go
